fix(cmd): add 'extensions' as a parent command
