--- pkg/cmd.go
@@ -1,46 +1,0 @@
-package pkg
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-
-	homedir "github.com/mitchellh/go-homedir"
-	"github.com/spf13/viper"
-)
-
-var cfgFile string
-
-var rootCmd = &cobra.Command{
-	Use:   "",
-	Short: "",
-	Long:  ``,
-	Run:   func(cmd *cobra.Command, args []string) {},
-}
-
-func Execute() {
-	cobra.CheckErr(rootCmd.Execute())
-}
-
-func init() {
-	cobra.OnInitialize(initConfig)
-	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.charmil.yaml)")
-	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
-
-func initConfig() {
-	if cfgFile != "" {
-		viper.SetConfigFile(cfgFile)
-	} else {
-		home, err := homedir.Dir()
-		cobra.CheckErr(err)
-		viper.AddConfigPath(home)
-		viper.SetConfigName(".charmil")
-	}
-
-	viper.AutomaticEnv()
-	if err := viper.ReadInConfig(); err == nil {
-		fmt.Fprintln(os.Stderr, "Using config file:", viper.ConfigFileUsed())
-	}
-}
