--- pkg/install.go
@@ -1,20 +1,0 @@
-package pkg
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-var installCmd = &cobra.Command{
-	Use:   "Install Extension",
-	Short: "",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("install called")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(installCmd)
-}
