--- pkg/list.go
@@ -1,20 +1,0 @@
-package pkg
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-var listCmd = &cobra.Command{
-	Use:   "List all Extensions",
-	Short: "",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("list called")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(listCmd)
-}
