--- pkg/remove.go
@@ -1,20 +1,0 @@
-package pkg
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-var removeCmd = &cobra.Command{
-	Use:   "Remove Extension",
-	Short: "",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("remove called")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(removeCmd)
-}
